+++ wai-extra/Network/Wai/Middleware/RequestLogger.hs
@@ -41,6 +41,7 @@ import Network.Wai.Internal (Response (..))
 import Data.Default.Class (Default (def))
 import Network.Wai.Logger
 import Network.Wai.Middleware.RequestLogger.Internal
+import Network.Wai.Header (contentLength)
 import Data.Text.Encoding (decodeUtf8')
 
 data OutputFormat = Apache IPAddrSource
@@ -106,11 +107,7 @@ mkRequestLogger RequestLoggerSettings{..} = do
 
 apacheMiddleware :: ApacheLoggerActions -> Middleware
 apacheMiddleware ala app req sendResponse = app req $ \res -> do
-    let msize = lookup H.hContentLength (responseHeaders res) >>= readInt'
-        readInt' bs =
-            case S8.readInteger bs of
-                Just (i, "") -> Just i
-                _ -> Nothing
+    let msize = contentLength (responseHeaders res)
     apacheLogger ala req (responseStatus res) msize
     sendResponse res
 
@@ -246,7 +243,7 @@ detailedMiddleware' :: Callback
                     -> (BS.ByteString -> BS.ByteString -> [BS.ByteString])
                     -> Middleware
 detailedMiddleware' cb ansiColor ansiMethod ansiStatusCode app req sendResponse = do
-    let mlen = lookup H.hContentLength (requestHeaders req) >>= readInt
+    let mlen = contentLength (requestHeaders req)
     (req', body) <-
         case mlen of
             -- log the request body if it is small
@@ -308,10 +305,7 @@ detailedMiddleware' cb ansiColor ansiMethod ansiStatusCode app req sendResponse 
     collectPostParams (postParams, files) = postParams ++
       map (\(k,v) -> (k, "FILE: " <> fileName v)) files
 
-    readInt bs =
-        case reads $ unpack bs of
-            (i, _):_ -> Just (i :: Int)
-            [] -> Nothing
+
 
 statusBS :: Response -> BS.ByteString
 statusBS = pack . show . statusCode . responseStatus