+++ wai-extra/ChangeLog.md
@@ -1,3 +1,8 @@
+## 3.0.12
+
+* Add Network.Wai.Header.contentLength to read the Content-Length header
+* The gzip middleware no longer zips responses smaller than 860 bytes
+
 ## 3.0.11
 
 * Add constructor for more detailed custom output formats for RequestLogger